fix: break static plugin import cycle
